+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make obsolete `ExceptionFactory.CreateForUnsupportedEnumValue` method.
 - Change `RetrySettings.DefaultInterval` value from 500 to 200 milliseconds.
 - Upgrade Selenium.WebDriver package reference to v4.36.0.
+- Add handling of `OpenQA.Selenium.UnknownErrorException` to `StaleSafely.Execute` methods.
 
 ### Removed